Enhance entity relationships and update property types
- Added new properties and navigation relationships in the `Amenity`, `Booking`, `Photo`, `Review`, and `User` classes.
- Changed data types from `long` to `int` for `UserId` and `RegionId` in multiple classes.
- Introduced new collections in `Landlord`, `PropertyType`, and `Tenant` for better data management.
- Initialized collections in `Permission`, `Role`, and `User` classes to ensure proper instantiation.

+++ src/RentalHome.Core/Entities/District.cs
@@ -1,9 +1,4 @@
 ﻿using RentalHome.Core.Common;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RentalHome.Core.Entities;
 
@@ -11,7 +6,7 @@ public  class District : BaseEntity
 {
     public string Name { get; set; }
 
-    public long RegionId { get; set; }
+    public int RegionId { get; set; }
 
     public Region Region { get; set; }
